feat: dedicated flags file
